+++ Cargo.toml
@@ -25,4 +25,3 @@ uuid = {version ="1.3.3", features = ["serde", "v4"]}
 chrono-tz = "0.8.3"
 serde_with= { version = "3.3.0", features = [] }
 ring = { version = "0.17.3", features = [] }
-gethostname = "0.3.0"
